Abstract Base class for DSM data
*Broke dsm data into a generic abstract class which will be inherited by the different types of dsms
*Will not build as the gui methods need to be refactored
