+++ src/gui/MatrixGuiHandler.java
@@ -1,8 +1,8 @@
 package gui;
 
-import DSMData.DSMConnection;
-import DSMData.DSMData;
-import DSMData.DSMItem;
+import Data.DSMConnection;
+import Data.SymmetricDSM;
+import Data.DSMItem;
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
@@ -212,7 +212,7 @@ public class MatrixGuiHandler {
 
     }
 
-    DSMData matrix;
+    SymmetricDSM matrix;
     public static final Background DEFAULT_BACKGROUND = new Background(new BackgroundFill(Color.color(1, 1, 1), new CornerRadii(3), new Insets(0)));
     public static final Background UNEDITABLE_CONNECTION_BACKGROUND = new Background(new BackgroundFill(Color.color(0, 0, 0), new CornerRadii(3), new Insets(0)));
     public static final Background HIGHLIGHT_BACKGROUND = new Background(new BackgroundFill(Color.color(.9, 1, 0), new CornerRadii(3), new Insets(0)));
@@ -235,10 +235,10 @@ public class MatrixGuiHandler {
     /**
      * Returns a MatrixGuiHandler object for a given matrix
      *
-     * @param matrix   the DSMData object to display
+     * @param matrix   the SymmetricDSM object to display
      * @param fontSize the default font size to display the matrix with
      */
-    public MatrixGuiHandler(DSMData matrix, double fontSize) {
+    public MatrixGuiHandler(SymmetricDSM matrix, double fontSize) {
         this.matrix = matrix;
         cells = new Vector<>();
         gridUidLookup = new HashMap<>();
@@ -438,7 +438,7 @@ public class MatrixGuiHandler {
 
 
     /**
-     * Creates the gui that displays a matrix. Uses the DSMData's getGridArray() method to create the grid.
+     * Creates the gui that displays a matrix. Uses the SymmetricDSM's getGridArray() method to create the grid.
      * Puts grid in a scroll pane and adds a location label (displays connection row, column) at the bottom of the VBox.
      * Returns the VBox so that it can be added to a layout
      *
